+++ app/src/main/java/com/shub39/rush/presentation/ListItemExt.kt
@@ -29,7 +29,9 @@ private const val endCornerRadius = 16
 
 @Composable
 fun listItemColors(): ListItemColors {
-    return ListItemDefaults.colors(containerColor = MaterialTheme.colorScheme.surfaceContainerHighest)
+    return ListItemDefaults.colors(
+        containerColor = MaterialTheme.colorScheme.surfaceContainerHighest
+    )
 }
 
 fun leadingItemShape(): Shape =